feat(ios): drop 4 legacy Mullvad source files from build targets (EphemeralPeerNegotiator + EphemeralPeerExchangeActor + EphemeralPeerReceiver + TunnelObfuscator + MullvadPostQuantum+Stubs) — Warren Quinn/wallet path doesn't use ephemeral-peer PQ or obfuscator proxy concepts. Files preserved on disk for git history (mirrors C.4.4.X WG pattern). WireGuardKey.swift KEPT (provides WireGuard.PublicKey namespace used by RESTTypes.Device, full rebrand deferred C.4.5). Fix WarrenWallet.swift payload.count Int → UInt cast (warren_wallet_sign FFI expects uintptr_t). Package.resolved updated (swift-log dropped via C.2.X.C vendoring). 2 helper scripts : remove-mullvad-legacy-sources.rb + restore-wireguardkey.rb idempotent rollback

+++ ios/WarrenRustRuntime/WarrenWallet.swift
@@ -138,7 +138,7 @@ public final class WarrenWallet {
                     warren_wallet_sign(
                         seedRaw.bindMemory(to: UInt8.self).baseAddress,
                         payloadRaw.bindMemory(to: UInt8.self).baseAddress,
-                        payload.count,
+                        UInt(payload.count),
                         sigPtr.baseAddress
                     )
                 }
